SadTalker_service/src/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# PDF İşleme
PyPDF2==3.0.1

# Sunum Oluşturma
python-pptx==0.6.23

# Yapay Zeka ve Google API
google-generativeai==0.3.0

# Web Scraping ve Görsel İndirme
selenium==4.16.0
webdriver-manager==4.0.1
Pillow==10.2.0
requests==2.31.0

# API ve Veri Doğrulama
pydantic==2.5.3

# SadTalker Gereksinimleri
torch==2.1.0
numpy==1.24.3
scipy==1.10.1
opencv-python==4.8.0.74
tqdm==4.65.0
face-alignment==1.3.5
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
gfpgan==1.3.8